third_party/blink/web_tests/platform/generic/editing/selection/range-insertNode-same-container-expected.txt

PASS range.toString() is "A"
PASS range.toString() is "A"
PASS range.toString() is "C1DEC2"
PASS range.toString() is "C1DEC2"
PASS range.toString() is "F"
PASS range.toString() is "C1DEC2"
PASS range.startContainer is containerElement
PASS range.startOffset is 4
PASS range.endContainer is containerElement
PASS range.endOffset is 4
PASS range.toString() is "C1DEC2"
PASS successfullyParsed is true

TEST COMPLETE

F
B
A
C1
D
E
C2
G
Test various cases of Range.insertNode with nodes in the same container as the Range.
